+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Other notable changes:
 * `net-util`:
   * Fixed `IncomingRequest.getHeaderOrNull()`, which was very broken.
   * Added `DispatchInfo.isFile()`.
+  * Added a handful of static getters to `StatusResponse`.
 
 ### v0.7.8 -- 2024-07-30 -- stable release